refactor: Update package.json and sanity.config.ts for simpler color input plugin
This commit updates the package.json file to add the "sanity-plugin-simpler-color-input" dependency with version "^2.2.2". It also updates the sanity.config.ts file to configure the "simplerColorInput" plugin with a default color format of "hex" and a default color list that includes a "Custom..." option and the colors defined in the COLORS module. These changes align with recent updates made to enhance the color input functionality and improve the user experience.