feat: Add addToWaitlist function to actions.ts
This commit adds the addToWaitlist function to the actions.ts file. The function allows users to add their email to the waitlist by creating a new waitlist object in the Sanity database. This addition enhances the functionality of the project by enabling users to join the waitlist for updates and notifications.